+++ services/api-server/tests/unit/api_solvers/test_api_routers_solvers_jobs_delete.py
@@ -11,7 +11,7 @@ import pytest
 from faker import Faker
 from models_library.basic_regex import UUID_RE_BASE
 from pydantic import parse_file_as
-from pytest_simcore.helpers.httpx_calls_capture_model import HttpApiCallCaptureModel
+from pytest_simcore.helpers.httpx_calls_capture_models import HttpApiCallCaptureModel
 from respx import MockRouter
 from simcore_service_api_server.models.schemas.jobs import Job, JobInputs
 from starlette import status
