+++ services/api-server/tests/unit/test_api_files.py
@@ -25,7 +25,7 @@ from models_library.api_schemas_storage import (
 )
 from models_library.basic_types import SHA256Str
 from pydantic import parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture_model import (
+from pytest_simcore.helpers.httpx_calls_capture_models import (
     CreateRespxMockCallback,
     HttpApiCallCaptureModel,
 )
