+++ services/api-server/tests/unit/test_api_solver_jobs.py
@@ -18,7 +18,7 @@ from models_library.api_schemas_webserver.resource_usage import PricingUnitGet
 from models_library.api_schemas_webserver.wallets import WalletGetWithAvailableCredits
 from models_library.generics import Envelope
 from pydantic import parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture_model import (
+from pytest_simcore.helpers.httpx_calls_capture_models import (
     CreateRespxMockCallback,
     HttpApiCallCaptureModel,
     SideEffectCallback,
