+++ services/api-server/tests/unit/test_api_solvers.py
@@ -11,7 +11,7 @@ from fastapi import status
 from httpx import AsyncClient
 from models_library.api_schemas_api_server.pricing_plans import ServicePricingPlanGet
 from pydantic import parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture_model import CreateRespxMockCallback
+from pytest_simcore.helpers.httpx_calls_capture_models import CreateRespxMockCallback
 from simcore_service_api_server._meta import API_VTAG
