+++ services/api-server/tests/unit/test_api_wallets.py
@@ -12,7 +12,7 @@ from fastapi import status
 from httpx import AsyncClient
 from models_library.api_schemas_webserver.wallets import WalletGetWithAvailableCredits
 from pydantic import parse_obj_as
-from pytest_simcore.helpers.httpx_calls_capture_model import (
+from pytest_simcore.helpers.httpx_calls_capture_models import (
     CreateRespxMockCallback,
     HttpApiCallCaptureModel,
 )
