+++ services/api-server/tests/unit/test_utils_http_calls_capture.py
@@ -14,7 +14,7 @@ import respx
 from faker import Faker
 from models_library.basic_regex import UUID_RE_BASE
 from pydantic import HttpUrl
-from pytest_simcore.helpers.httpx_calls_capture_model import HttpApiCallCaptureModel
+from pytest_simcore.helpers.httpx_calls_capture_models import HttpApiCallCaptureModel
 
 
 async def test_capture_http_call(
